+++ src/main/java/com/robertx22/age_of_exile/database/data/skill_gem/SkillGemTag.java
@@ -1,0 +1,12 @@
+package com.robertx22.age_of_exile.database.data.skill_gem;
+
+public enum SkillGemTag {
+
+    PROJECTILE,
+    DAMAGE,
+    HEALING,
+    AREA,
+    AURA,
+    TOTEM,
+
+}